Restructured Components Grouped together Auth components and Shared components
