+++ src/components/Auth/SignIn/SignIn.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react'
 import { withRouter } from 'react-router-dom'
 
-import { signIn } from '../../api/auth'
-import messages from '../AutoDismissAlert/messages'
+import { signIn } from '../../../api/auth'
+import messages from '../../Shared/AutoDismissAlert/messages'
 
 import Form from 'react-bootstrap/Form'
 import Button from 'react-bootstrap/Button'
