+++ src/components/Auth/SignOut/SignOut.js
@@ -1,8 +1,8 @@
 import { Component } from 'react'
 import { withRouter } from 'react-router-dom'
 
-import { signOut } from '../../api/auth'
-import messages from '../AutoDismissAlert/messages'
+import { signOut } from '../../../api/auth'
+import messages from '../../Shared/AutoDismissAlert/messages'
 
 class SignOut extends Component {
   componentDidMount () {
